move/#122: drawable res 분리
