--- core/designsystem/src/main/res/drawable/ic_appbar_add.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="14dp"
-    android:height="14dp"
-    android:viewportWidth="14"
-    android:viewportHeight="14">
-  <path
-      android:pathData="M7,0C7.552,0 8,0.448 8,1V6H13C13.552,6 14,6.448 14,7C14,7.552 13.552,8 13,8H8V13C8,13.552 7.552,14 7,14C6.448,14 6,13.552 6,13V8H1C0.448,8 0,7.552 0,7C0,6.448 0.448,6 1,6H6V1C6,0.448 6.448,0 7,0Z"
-      android:fillColor="#959595"
-      android:fillType="evenOdd"/>
-</vector>
