--- core/designsystem/src/main/res/drawable/ic_appbar_setting.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="18dp"
-    android:height="18dp"
-    android:viewportWidth="18"
-    android:viewportHeight="18">
-  <path
-      android:pathData="M10.49,2.17C10.11,0.61 7.89,0.61 7.51,2.17C7.453,2.404 7.342,2.622 7.185,2.805C7.028,2.989 6.83,3.132 6.608,3.225C6.385,3.317 6.143,3.356 5.903,3.337C5.662,3.319 5.43,3.244 5.224,3.118C3.852,2.282 2.282,3.852 3.118,5.224C3.658,6.11 3.179,7.266 2.171,7.511C0.61,7.89 0.61,10.111 2.171,10.489C2.405,10.546 2.623,10.657 2.806,10.814C2.99,10.971 3.133,11.169 3.226,11.392C3.318,11.615 3.356,11.856 3.337,12.097C3.319,12.337 3.243,12.57 3.118,12.776C2.282,14.148 3.852,15.718 5.224,14.882C5.43,14.757 5.662,14.681 5.903,14.663C6.144,14.644 6.385,14.682 6.608,14.774C6.831,14.867 7.029,15.01 7.186,15.194C7.343,15.377 7.454,15.595 7.511,15.829C7.89,17.39 10.111,17.39 10.489,15.829C10.546,15.595 10.657,15.377 10.814,15.194C10.971,15.011 11.169,14.867 11.392,14.775C11.615,14.683 11.856,14.644 12.097,14.663C12.337,14.682 12.57,14.757 12.776,14.882C14.148,15.718 15.718,14.148 14.882,12.776C14.757,12.57 14.681,12.337 14.663,12.097C14.644,11.857 14.683,11.615 14.775,11.392C14.867,11.169 15.011,10.972 15.194,10.815C15.377,10.658 15.595,10.546 15.829,10.489C17.39,10.11 17.39,7.889 15.829,7.511C15.594,7.454 15.377,7.343 15.193,7.186C15.01,7.029 14.866,6.831 14.774,6.608C14.682,6.385 14.644,6.144 14.662,5.903C14.681,5.663 14.756,5.43 14.882,5.224C15.718,3.852 14.148,2.282 12.776,3.118C12.57,3.244 12.337,3.319 12.097,3.338C11.856,3.356 11.615,3.318 11.392,3.226C11.169,3.133 10.971,2.99 10.814,2.807C10.657,2.623 10.546,2.405 10.489,2.171L10.49,2.17ZM9,12C9.796,12 10.559,11.684 11.121,11.122C11.684,10.559 12,9.796 12,9C12,8.204 11.684,7.441 11.121,6.879C10.559,6.316 9.796,6 9,6C8.204,6 7.441,6.316 6.879,6.879C6.316,7.441 6,8.204 6,9C6,9.796 6.316,10.559 6.879,11.122C7.441,11.684 8.204,12 9,12V12Z"
-      android:strokeWidth="1.5"
-      android:fillColor="#00000000"
-      android:fillType="evenOdd"
-      android:strokeColor="#959595"/>
-</vector>
