--- core/designsystem/src/main/res/drawable/ic_star.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="14dp"
-    android:height="15dp"
-    android:viewportWidth="14"
-    android:viewportHeight="15">
-  <path
-      android:pathData="M13.547,5.43L9.351,4.791L7.474,0.808C7.423,0.699 7.339,0.611 7.235,0.557C6.974,0.422 6.656,0.535 6.526,0.808L4.649,4.791L0.453,5.43C0.337,5.447 0.231,5.504 0.15,5.591C0.052,5.696 -0.002,5.838 0,5.985C0.002,6.132 0.059,6.272 0.16,6.375L3.197,9.475L2.479,13.853C2.462,13.955 2.473,14.059 2.51,14.155C2.547,14.251 2.609,14.334 2.689,14.394C2.769,14.455 2.864,14.491 2.962,14.499C3.06,14.506 3.159,14.484 3.246,14.436L7,12.369L10.754,14.436C10.856,14.493 10.976,14.512 11.09,14.491C11.377,14.439 11.571,14.154 11.521,13.853L10.804,9.475L13.84,6.375C13.923,6.29 13.977,6.179 13.994,6.058C14.038,5.755 13.837,5.475 13.547,5.43Z"
-      android:fillColor="#346CFD"/>
-</vector>
