+++ core/ui/src/main/res/drawable/ic_align_checked.xml
