+++ core/ui/src/main/res/drawable/ic_arrow_gray_right.xml
