+++ core/ui/src/main/res/drawable/ic_star_empty.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="12dp"
+    android:height="12dp"
+    android:viewportWidth="12"
+    android:viewportHeight="12">
+  <path
+      android:pathData="M10.911,4.447L7.763,3.968L6.356,0.981C6.317,0.899 6.254,0.833 6.176,0.793C5.98,0.692 5.742,0.776 5.644,0.981L4.237,3.968L1.089,4.447C1.003,4.46 0.923,4.503 0.863,4.568C0.789,4.647 0.749,4.753 0.75,4.864C0.751,4.974 0.795,5.079 0.87,5.156L3.147,7.481L2.609,10.764C2.597,10.841 2.605,10.919 2.633,10.991C2.66,11.063 2.707,11.125 2.767,11.171C2.827,11.216 2.898,11.243 2.971,11.249C3.045,11.255 3.119,11.238 3.185,11.202L6,9.652L8.815,11.202C8.892,11.245 8.982,11.259 9.067,11.244C9.283,11.205 9.428,10.99 9.391,10.764L8.853,7.481L11.13,5.156C11.192,5.093 11.233,5.009 11.245,4.919C11.279,4.691 11.128,4.481 10.911,4.447Z"
+      android:fillColor="#DADADA"/>
+</vector>
