+++ feature/lectureevaluation/viewerreporter/src/main/java/com/suwiki/feature/lectureevaluation/viewerreporter/component/EvaluationSearchBar.kt
@@ -30,7 +30,6 @@ import androidx.compose.ui.platform.SoftwareKeyboardController
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.component.searchbar.BasicSearchBar
 import com.suwiki.core.designsystem.shadow.cardShadow
 import com.suwiki.core.designsystem.theme.Gray6A
@@ -39,6 +38,7 @@ import com.suwiki.core.designsystem.theme.GrayFB
 import com.suwiki.core.designsystem.theme.SuwikiTheme
 import com.suwiki.core.designsystem.theme.White
 import com.suwiki.core.ui.extension.suwikiClickable
+import com.suwiki.feature.lectureevaluation.viewerreporter.R
 
 @Composable
 fun EvaluationSearchBar(
