+++ feature/lectureevaluation/viewerreporter/src/main/java/com/suwiki/feature/lectureevaluation/viewerreporter/detail/component/ExamEvaluationContainer.kt
@@ -15,7 +15,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.component.badge.BadgeColor
 import com.suwiki.core.designsystem.component.badge.SuwikiBadge
 import com.suwiki.core.designsystem.component.button.SuwikiContainedSmallButton
@@ -35,7 +34,7 @@ fun ExamEvaluationContainer(
   isAuthor: Boolean = false,
   onClickButton: () -> Unit,
 ) {
-  val buttonText = if (isAuthor) stringResource(id = R.string.word_edit) else stringResource(id = R.string.word_report)
+  val buttonText = if (isAuthor) stringResource(id = com.suwiki.core.ui.R.string.word_edit) else stringResource(id = com.suwiki.core.ui.R.string.word_report)
 
   Column(
     modifier = Modifier
@@ -49,7 +48,7 @@ fun ExamEvaluationContainer(
       if (isAuthor) {
         SuwikiBadge(
           color = BadgeColor.Blue,
-          text = stringResource(id = R.string.word_my),
+          text = stringResource(id = com.suwiki.core.ui.R.string.word_my),
         )
         Spacer(modifier = Modifier.width(8.dp))
       }
@@ -71,7 +70,7 @@ fun ExamEvaluationContainer(
       horizontalArrangement = Arrangement.spacedBy(6.dp),
     ) {
       Text(
-        text = stringResource(id = R.string.word_difficulty),
+        text = stringResource(id = com.suwiki.core.ui.R.string.word_difficulty),
         style = SuwikiTheme.typography.caption2,
         color = Gray95,
       )
@@ -87,7 +86,7 @@ fun ExamEvaluationContainer(
       horizontalArrangement = Arrangement.spacedBy(6.dp),
     ) {
       Text(
-        text = stringResource(id = R.string.word_type_exam),
+        text = stringResource(id = com.suwiki.core.ui.R.string.word_type_exam),
         style = SuwikiTheme.typography.caption2,
         color = Gray95,
       )
