+++ feature/lectureevaluation/viewerreporter/src/main/res/drawable/ic_filter.xml
