+++ feature/lectureevaluation/viewerreporter/src/main/res/drawable/ic_filter_arrow_down.xml
