+++ feature/openmajor/src/main/java/com/suwiki/feature/openmajor/component/OpenMajorContainer.kt
@@ -20,7 +20,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.suwiki.core.designsystem.R
 import com.suwiki.core.designsystem.theme.Black
 import com.suwiki.core.designsystem.theme.Blue5
 import com.suwiki.core.designsystem.theme.GrayDA
@@ -28,6 +27,7 @@ import com.suwiki.core.designsystem.theme.Primary
 import com.suwiki.core.designsystem.theme.SuwikiTheme
 import com.suwiki.core.designsystem.theme.White
 import com.suwiki.core.ui.extension.suwikiClickable
+import com.suwiki.feature.openmajor.R
 
 @Composable
 fun OpenMajorContainer(
