+++ feature/openmajor/src/main/res/drawable/ic_filter_stared.xml
