+++ feature/timetable/src/main/java/com/suwiki/feature/timetable/widget/timetable/Common.kt
@@ -3,7 +3,7 @@ package com.suwiki.feature.timetable.widget.timetable
 import android.content.Context
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
-import com.suwiki.core.designsystem.R
+import com.suwiki.core.ui.R
 import com.suwiki.core.model.timetable.TimetableDay
 
 internal val glanceTimetableHeightPerHour = 48.dp
